ca-gpt-free-v9/ai-service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "1.0.0"
description = "AI Service for CA-GPT-Free v9"
authors = ["AIGIS v1.0"]
packages = [{include = "ai_service"}, {include = "shared"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
PyJWT = "^2.8.0"
motor = "^3.4.0"
ruamel-yaml = "^0.18.6"
httpx = "^0.27.0"
tenacity = "^8.2.3"
beanie = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
mypy = "^1.10.0"
black = "^24.4.2"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
